requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
PIL==1.1.7
Pillow==1.7.8
Werkzeug==0.9.4
colormath==1.0.9
-e git+https://github.com/jyotiska/colorweave.git@f7af9edfd025de83e2f8003090f641735d97f5fb#egg=colorweave
ecdsa==0.10
httplib2==0.8
itsdangerous==0.23
numpy==1.8.1
oauthlib==0.6.0
paramiko==1.12.1
pycrypto==2.6.1
python-instagram==0.8.0
requests==2.1.0
requests-oauthlib==0.4.0
simplejson==3.3.2
ssh==1.8.0
twython==3.1.2
webcolors==1.4
wsgiref==0.1.2
